approve document

+++ music/src/main/java/com/boot/music/repositories/DocumentRepo.java
@@ -1,11 +1,9 @@
 package com.boot.music.repositories;
 
 import java.util.List;
-
+import com.boot.music.entity.Document;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.boot.music.entity.Document;
-
-public interface DocumentRepo extends JpaRepository<Document, Long>{
+public interface DocumentRepo extends JpaRepository<Document, Integer> {
 	List<Document> findByStatusID(short statusID);
 }
